+++ src/it/scala/au/com/agl/arc/load/KafkaLoadSuite.scala
@@ -13,6 +13,7 @@ import org.apache.commons.io.FileUtils
 import org.apache.commons.io.IOUtils
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
+import org.apache.spark.sql.types._
 
 import org.apache.kafka.clients.consumer.KafkaConsumer
 import org.apache.kafka.clients.consumer.ConsumerConfig
@@ -28,7 +29,7 @@ class KafkaLoadSuite extends FunSuite with BeforeAndAfter {
   val inputView = "inputView"
   val outputView = "outputView"
   val bootstrapServers = "localhost:29092"
-  val timeout = 3000L
+  val timeout = 1000L
   val checkPointPath = "/tmp/checkpoint"
 
   before {
@@ -65,6 +66,8 @@ class KafkaLoadSuite extends FunSuite with BeforeAndAfter {
       .withColumn("normal", randn(seed=27))
       .repartition(10)
       .toJSON
+      .select(col("value").cast(BinaryType))
+
     dataset.createOrReplaceTempView(inputView)
 
     load.KafkaLoad.load(
@@ -96,12 +99,12 @@ class KafkaLoadSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty 
       )
     ).get 
 
     val expected = dataset
-    val actual = extractDataset.select($"value").as[String]
+    val actual = extractDataset.select($"value")
 
     val actualExceptExpectedCount = actual.except(expected).count
     val expectedExceptActualCount = expected.except(actual).count
@@ -133,8 +136,9 @@ class KafkaLoadSuite extends FunSuite with BeforeAndAfter {
       .withColumn("uniform", rand(seed=10))
       .withColumn("normal", randn(seed=27))
       .withColumn("value", to_json(struct($"uniform", $"normal")))
-      .select("key", "value")
+      .select(col("key").cast(BinaryType), col("value").cast(BinaryType))
       .repartition(10)
+
     dataset.createOrReplaceTempView(inputView)
 
     load.KafkaLoad.load(
@@ -166,7 +170,7 @@ class KafkaLoadSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty
+        params=Map.empty 
       )
     ).get